==> code/02_preprocessing.ipynb <==
# Notebook 02: Preprocessing

Preprocessing the text is done here before EDA and modeling. Four functions are developed to do the following tasks: lowercasing, removing punctuation, removing stop words and tokenizing.

import pandas as pd
pd.set_option('display.max_rows', None)
pd.set_option('display.max_columns', None)
pd.set_option('display.width', None)
pd.set_option('display.max_colwidth', None)
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

from nltk.tokenize import word_tokenize
from gensim.utils import simple_preprocess
from nltk.corpus import stopwords

df_all = pd.read_csv('./sutta_csv/cleaned/df_all_clean.csv')

#### Lowercasing

def lowercase(df):
    df['lower_text'] = df['text_full'].str.lower().copy()
    return df

#### Removing punctuation

def remove_punct(df):
    df['no_punct'] = df['lower_text'].str.replace('.', ' ').copy()
    df['no_punct'] = df['lower_text'].str.replace('--', ' ').copy()
    df['no_punct'] = df['lower_text'].str.replace('-', ' ').copy()
    df['no_punct'] = df['lower_text'].str.replace('?', ' ').copy()
    df['no_punct'] = df['lower_text'].str.replace('(', ' ').copy()
    df['no_punct'] = df['lower_text'].str.replace(')', ' ').copy()
    df['no_punct'] = df['lower_text'].str.replace(',', '').copy()
    df['no_punct'] = df['lower_text'].str.replace('  ', ' ').copy()
    df['no_punct'] = df['lower_text'].str.replace('[^\w\s]', "").copy()
    return df

#### Tokenizing

def token(df):
    df['text_token'] = [simple_preprocess(each) for each in df['no_punct']]
    return df

#### Removing Stop Words

my_stopwords = stopwords.words('english')

#Base code from: https://www.analyticsvidhya.com/blog/2018/10/stepwise-guide-topic-modeling-latent-semantic-analysis/
def stopwords(df):
    df['text_token_no_stop'] = df['text_token'].apply(lambda x: [item for item in x if item not in my_stopwords])
    df['text_no_stop'] = df['no_punct'].apply(lambda x: [item for item in x if item not in my_stopwords])
    return df

#### Apply all functions

def preprocess(df):
    df = stopwords(token(remove_punct(lowercase(df))))
    return df

df_all = preprocess(df_all);

# df_all.to_csv('./sutta_csv/cleaned/df_all_prep.csv', index= False)
